Warn if Tailscale isn't installed #66
 * This adds a pre-bootstrap check to see if the tailscale binary is installed/callable
 * If not, it will warn and close down
 * If it's present and callable, we'll bootstrap and run
 * Fixed a few shutdown issues/nullref checks needed

+++ src/IpnWatcher.cpp
@@ -41,6 +41,10 @@ void IpnWatcher::start() {
 
 void IpnWatcher::stop() {
     m_isShuttingDown = true;
+    if (m_process == nullptr) {
+        return;
+    }
+
     if (m_process->state() == QProcess::Running) {
         m_process->terminate();
     }
